Ability to mark songs unplayable.
There are multiple situations where songs may not be
playable. They may not be found (really bad data, or
rare song), or they may be found, but not playable to
due legal reasons. As a result, we need a mechanism
that can prevent unplayable songs from reaching the
queue, where users may get hopes about an upcoming
song that will never be played.

This implementation is purely transient; when the
server restarts, it must learn what songs are unplayable.